Fix "Unknown Error" on Playback Request Due to InnerTube API Changes (#468)
* ## Fix "Unknown Error" on Playback Request Due to InnerTube API Changes

- Replaced Piped Streams with direct YouTube streams for `YouTube.player()`, as [pipedapi.kevin.rocks](https://pipedapi.kevin.rocks) is down.
  **Note**: Direct streams seem to cause issues once in a while, such as playback interrupts or timeouts.
- Added support for signature timestamps in `YouTube.player()`, fetched from InnerTube using the media ID, for clients requiring signature timestamps.
- Playback and stream fetching handled in `innertune/utils/YTPlayerUtils.kt`.
- Extraction/Deobfuscation of `PlayerResponse` data is handled in `innertube/utils/NewPipeUtils.kt`.
- Updated 'YoutubeClient' clients to reflect changes.
- Fixed Kugou test causing build errors

### Additional Changes:
- Added [libs.newpipeExtractor](https://github.com/TeamNewPipe/NewPipeExtractor) to `libs.versions.toml` and `innertube/build.gradle`.

* Should fix downloads from being throttled, and removes api key from player requests (doesn't appear to be used anymore)

+++ innertube/src/main/java/com/malopieds/innertube/models/body/PlayerBody.kt
@@ -9,7 +9,7 @@ data class PlayerBody(
     val videoId: String,
     val playlistId: String?,
     val cpn: String? = "wzf9Y0nqz6AUe2Vr", // need some random cpn to get same algorithm for sig
-    val playbackContext: PlaybackContext? = PlaybackContext(ContentPlaybackContext(20052L)),
+    val playbackContext: PlaybackContext? = PlaybackContext(ContentPlaybackContext(20052)),
 ) {
     @Serializable
     data class PlaybackContext(
@@ -18,6 +18,6 @@ data class PlayerBody(
 
     @Serializable
     data class ContentPlaybackContext(
-        val signatureTimestamp: Long?,
+        val signatureTimestamp: Int?,
     )
 }
